requirements.txt
pytube==12.1.0
telebot==0.0.5
pyTelegramBotAPI==4.7.0
wheel==0.37.1
pillow-heif==0.7.0
Pillow==9.2.0
rembg==2.0.25
pathlib==1.0.1